Refactor process widgets
